refactor: move aggregates to model folder
