+++ src/renderer/widgets/UnlockModal/index.ts
@@ -1,4 +1,4 @@
 export { UnlockModal } from './ui/UnlockModal';
 export { UnlockConfirmation } from './ui/UnlockConfirmation';
-export { unlockAggregate } from './aggregates/unlock';
-export { unlockConfirmAggregate } from './aggregates/unlockConfirm';
+export { unlockAggregate } from './model/unlock';
+export { unlockConfirmAggregate } from './model/unlockConfirm';
